+++ activerecord/CHANGELOG.md
@@ -1,3 +1,19 @@
+*   Use the right column to type cast grouped calculations with custom expressions.
+
+    Fixes #13230.
+
+    Example:
+
+            # Before
+            Account.group(:firm_name).sum('0.01 * credit_limit')
+            # => { '37signals' => '0.5' }
+
+            # After
+            Account.group(:firm_name).sum('0.01 * credit_limit')
+            # => { '37signals' => 0.5 }
+
+    *Paul Nikitochkin*
+
 *   Support deprecated finder options when performing calculations
     on an association.